+++ src/features/admin/serviceApiSlice.ts
@@ -98,6 +98,12 @@ export const apiSlice = createApi({
       query: () =>
         `/api/administration/serviceaccount/user/roles?languageShortName=${i18next.language}`,
     }),
+    resetCredential: builder.mutation<ServiceAccountDetail, string>({
+      query: (id: string) => ({
+        url: `api/administration/serviceaccount/owncompany/serviceaccounts/${id}/resetCredentials`,
+        method: 'POST',
+      }),
+    }),
   }),
 })
 
@@ -107,4 +113,5 @@ export const {
   useFetchServiceAccountListQuery,
   useFetchServiceAccountDetailQuery,
   useFetchServiceAccountRolesQuery,
+  useResetCredentialMutation,
 } = apiSlice